Deleted the unneeded 'MessageBusAgent' class (#240)
With only one MQTT agent, there is no need for an abstract base class.
