+++ src/main/scala/org/clulab/asist/agents/IdcWorker.scala
@@ -14,7 +14,7 @@ import scala.util.{Failure, Success}
  */
 
 class IdcWorker(
-  val owner: MessageBusAgent
+  val owner: DialogAgentMqtt
 ) extends LazyLogging {
 
   logger.info("IDC Worker ready.")
